+++ magik-language-server/src/test/java/nl/ramsolutions/sw/magik/languageserver/inlayhint/InlayHintProviderTest.java
@@ -62,6 +62,7 @@ class InlayHintProviderTest {
                     ParameterDefinition.Modifier.NONE,
                     TypeString.UNDEFINED)),
             null,
+            Collections.emptySet(),
             ExpressionResultString.UNDEFINED,
             ExpressionResultString.EMPTY));