+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/definitions/parsers/DefineSharedConstantParser.java
@@ -146,6 +146,7 @@ public class DefineSharedConstantParser {
             modifiers,
             parameters,
             null,
+            Collections.emptySet(),
             new ExpressionResultString(typeRef),
             ExpressionResultString.EMPTY);
     return List.of(methodDefinition);