+++ magik-typed-checks/src/main/java/nl/ramsolutions/sw/magik/typedchecks/CheckList.java
@@ -6,6 +6,8 @@ import nl.ramsolutions.sw.magik.checks.DisabledByDefault;
 import nl.ramsolutions.sw.magik.checks.MagikCheck;
 import nl.ramsolutions.sw.magik.checks.checks.TypeDocCheck;
 import nl.ramsolutions.sw.magik.typedchecks.checks.ConditionalExpressionIsFalseTypedCheck;
+import nl.ramsolutions.sw.magik.typedchecks.checks.DeprecatedMethodUsageTypedCheck;
+import nl.ramsolutions.sw.magik.typedchecks.checks.DeprecatedTypeUsageTypedCheck;
 import nl.ramsolutions.sw.magik.typedchecks.checks.GlobalExistsTypedCheck;
 import nl.ramsolutions.sw.magik.typedchecks.checks.MethodArgumentCountMatchesParameterCountTypedCheck;
 import nl.ramsolutions.sw.magik.typedchecks.checks.MethodArgumentTypeMatchesParameterTypeTypedCheck;
@@ -32,6 +34,8 @@ public final class CheckList {
   public static List<Class<? extends MagikCheck>> getChecks() {
     return List.of(
         ConditionalExpressionIsFalseTypedCheck.class,
+        DeprecatedMethodUsageTypedCheck.class,
+        DeprecatedTypeUsageTypedCheck.class,
         GlobalExistsTypedCheck.class,
         MethodArgumentCountMatchesParameterCountTypedCheck.class,
         MethodArgumentTypeMatchesParameterTypeTypedCheck.class,