+++ magik-typed-checks/src/test/java/nl/ramsolutions/sw/magik/typedchecks/checks/DeprecatedMethodUsageTypedCheckTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Collections;
 import java.util.EnumSet;
 import java.util.List;
+import java.util.Set;
 import nl.ramsolutions.sw.magik.analysis.definitions.DefinitionKeeper;
 import nl.ramsolutions.sw.magik.analysis.definitions.IDefinitionKeeper;
 import nl.ramsolutions.sw.magik.analysis.definitions.MethodDefinition;
@@ -14,22 +15,11 @@ import nl.ramsolutions.sw.magik.checks.MagikIssue;
 import nl.ramsolutions.sw.magik.typedchecks.MagikTypedCheck;
 import org.junit.jupiter.api.Test;
 
-/** Test MethodExistsTypedCheck. */
-class MethodExistsTypedCheckTest extends MagikTypedCheckTestBase {
+/** Test {@link DeprecatedMethodUsageTypedCheck}. */
+public class DeprecatedMethodUsageTypedCheckTest extends MagikTypedCheckTestBase {
 
-  @Test
-  void testMethodUnknown() {
-    final String code = "" + "_block\n" + "  object.m()\n" + "_endblock";
-    final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
-    final MagikTypedCheck check = new MethodExistsTypedCheck();
-    final List<MagikIssue> issues = this.runCheck(code, definitionKeeper, check);
-    assertThat(issues).hasSize(1);
-  }
-
-  @Test
-  void testMethodKnown() {
-    final String code = "" + "_block\n" + "  object.m()\n" + "_endblock";
-    final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
+  private void addMethodDefinition(
+      final IDefinitionKeeper definitionKeeper, final String... topics) {
     definitionKeeper.add(
         new MethodDefinition(
             null,
@@ -41,10 +31,33 @@ class MethodExistsTypedCheckTest extends MagikTypedCheckTestBase {
             EnumSet.noneOf(MethodDefinition.Modifier.class),
             Collections.emptyList(),
             null,
-            Collections.emptySet(),
-            new ExpressionResultString(TypeString.SW_OBJECT),
+            Set.of(topics),
+            ExpressionResultString.EMPTY,
             ExpressionResultString.EMPTY));
-    final MagikTypedCheck check = new MethodExistsTypedCheck();
+  }
+
+  @Test
+  void testMethodDeprecated() {
+    final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
+    this.addMethodDefinition(definitionKeeper, "deprecated");
+    final String code = """
+        _block
+          object.m()
+        _endblock""";
+    final MagikTypedCheck check = new DeprecatedMethodUsageTypedCheck();
+    final List<MagikIssue> issues = this.runCheck(code, definitionKeeper, check);
+    assertThat(issues).hasSize(1);
+  }
+
+  @Test
+  void testMethodNotDeprecated() {
+    final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
+    this.addMethodDefinition(definitionKeeper, "not_deprecated");
+    final String code = """
+        _block
+          object.m()
+        _endblock""";
+    final MagikTypedCheck check = new DeprecatedMethodUsageTypedCheck();
     final List<MagikIssue> issues = this.runCheck(code, definitionKeeper, check);
     assertThat(issues).isEmpty();
   }