+++ magik-typed-checks/src/test/java/nl/ramsolutions/sw/magik/typedchecks/checks/MethodArgumentTypeMatchesParameterTypeTypedCheckTest.java
@@ -2,6 +2,7 @@ package nl.ramsolutions.sw.magik.typedchecks.checks;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.Collections;
 import java.util.EnumSet;
 import java.util.List;
 import nl.ramsolutions.sw.magik.analysis.definitions.DefinitionKeeper;
@@ -39,6 +40,7 @@ class MethodArgumentTypeMatchesParameterTypeTypedCheckTest extends MagikTypedChe
                     ParameterDefinition.Modifier.NONE,
                     TypeString.SW_SYMBOL)),
             null,
+            Collections.emptySet(),
             ExpressionResultString.UNDEFINED,
             ExpressionResultString.EMPTY));
   }